Fix formatting. Fix mistaken declaration.

+++ src/openrct2/audio/Audio.cpp
@@ -228,7 +228,8 @@ AudioParams audio_get_params_from_location(RCT2Sound soundId, const LocationXYZ1
             int16_t vy = pos2.y - viewport->view_y;
             int16_t vx = pos2.x - viewport->view_x;
             params.pan = viewport->x + (vx >> viewport->zoom);
-            params.volume = SoundVolumeAdjust[static_cast<uint32_t>(soundId)] + ((-1024 * viewport->zoom - 1) * (1 << volumeDown)) + 1;
+            params.volume = SoundVolumeAdjust[static_cast<uint8_t>(soundId)]
+                + ((-1024 * viewport->zoom - 1) * (1 << volumeDown)) + 1;
 
             if (vy < 0 || vy >= viewport->view_height || vx < 0 || vx >= viewport->view_width || params.volume < -10000)
             {
